Changed Requests to a Request Session
I wasn't able to find a way to logout and invalidate the steamLoginSecure cookie using requests that were not part of a session. I rewrote the requests to be part of a singular session that is cached between program runs. The logout function now works and invalidates the steamLoginSecure cookie from being able to be used.