Added Print Response Function and Refactored Update Functions
Created a function that prints the response to a request when an error occurs. Also, refactored the update cards and packs functions to open database connection using with statements, made the update packs function retry when a total count of 0 is received, limited the retry attempts to 10, and changed updating messages.